Revert "refactor: breadcrumbs"
This reverts commit ffe2809fa6f64e510c27f0bf1a536dd663253a8a.
