--- src/theme/DocItem/styles.scss
@@ -1,36 +1,0 @@
-@use '../../css/components/mixins.scss';
-
-.breadcrumbs {
-  display: flex;
-  list-style: none;
-  margin: 0 0 2rem;
-  padding: 0;
-
-  > li {
-    display: flex;
-    align-items: center;
-
-    &:last-child {
-      color: var(--ifm-color-primary);
-      font-weight: 600;
-    }
-
-    &.separator {
-      color: rgba(0, 0, 0, 0.3);
-      margin: 0 10px;
-    }
-    &.category {
-      color: rgba(0, 0, 0, 0.5);
-      margin-right: 10px;
-    }
-
-    a {
-      color: black;
-
-      &:hover {
-        color: var(--ifm-link-color);
-        text-decoration: none;
-      }
-    }
-  }
-}
