+++ src/theme/Navbar/styles.scss
@@ -4,7 +4,7 @@
   --ifm-navbar-height: 6rem;
   --ifm-navbar-shadow: 0px 2px 18px 0px rgba(0, 0, 0, 0.04);
   --xtm-site-navbar-display: none;
-  --xtm-site-navbar-height: 0;
+  --xtm-site-navbar-height: 1.875rem;
 }
 
 @media (min-width: 997px) {
@@ -307,5 +307,141 @@
     margin-top: calc(
       -1 * (var(--ifm-navbar-height) + (var(--xtm-site-navbar-height) + 1px))
     ) !important;
+  }
+}
+
+.site-navbar {
+  position: sticky;
+  top: var(--ifm-navbar-height);
+  width: 100%;
+  display: var(--xtm-site-navbar-display);
+  background: var(--ifm-navbar-background-color);
+  font-size: 0.75rem;
+  font-weight: 600;
+  padding: 0 var(--ifm-navbar-padding-horizontal);
+  border-top: 1px solid rgba(0, 0, 0, 0.04);
+  z-index: calc(var(--ifm-z-index-fixed) + 10);
+  box-shadow: 0px 2px 18px rgba(0, 0, 0, 0.04);
+
+  &__inner {
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+  }
+
+  &__breadcrumbs,
+  &__secondary {
+    height: var(--xtm-site-navbar-height);
+    display: flex;
+    list-style: none;
+    margin: 0;
+    padding: 0;
+  }
+
+  &__breadcrumbs {
+    > li {
+      display: flex;
+      align-items: center;
+
+      &.separator {
+        color: rgba(0, 0, 0, 0.3);
+        margin: 0 10px;
+      }
+      &.category {
+        color: rgba(0, 0, 0, 0.5);
+        margin-right: 10px;
+      }
+
+      a {
+        color: black;
+
+        &:hover {
+          color: var(--ifm-link-color);
+          text-decoration: none;
+        }
+      }
+    }
+  }
+
+  &__item {
+    display: flex;
+    height: 100%;
+    align-items: center;
+    padding: 0 calc(var(--ifm-navbar-padding-horizontal) / 2);
+
+    a {
+      color: var(--ifm-color-black);
+      opacity: 0.5;
+      text-decoration: none;
+      transition-property: opacity, color, fill;
+      transition-duration: var(--ifm-transition-fast);
+      transition-timing-function: var(--ifm-transition-timing-default);
+
+      &:hover {
+        opacity: 1;
+      }
+    }
+
+    &--return {
+      display: flex;
+      padding-left: 0;
+      padding-right: var(--ifm-navbar-padding-horizontal);
+      margin-right: calc(var(--ifm-navbar-padding-horizontal) / 2);
+      box-shadow: 1px 0px 0px rgba(0, 0, 0, 0.04);
+
+      a {
+        opacity: 1;
+
+        &:hover {
+          color: var(--ifm-color-primary);
+        }
+
+        &::before {
+          content: '«';
+          color: var(--ifm-color-primary);
+          margin-right: 0.2rem;
+        }
+      }
+    }
+
+    &--contact {
+      padding-right: 0;
+
+      a {
+        display: flex;
+        align-items: center;
+        opacity: 1;
+        color: var(--ifm-color-primary);
+
+        &:hover {
+          color: var(--ifm-color-primary-darkest);
+
+          svg {
+            fill: var(--ifm-color-primary-darkest);
+          }
+        }
+
+        svg {
+          margin-right: 0.5rem;
+        }
+      }
+    }
+
+    &.dropdown--hoverable {
+      position: relative;
+
+      &:after {
+        @include mixins.down-arrow(var(--ifm-color-secondary-darker), 4px, 3px);
+        cursor: pointer;
+        margin: 2px 0 0 0.5rem;
+        content: '';
+      }
+    }
+  }
+
+  &__dropdown {
+    &__item {
+      padding: 0.5rem;
+    }
   }
 }
